pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dca-simulation"
version = "0.1.0"
description = "Execution Algorithm Simulator for DCA strategies"
authors = [{name = "DCA Simulator", email = "dev@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "simpy>=4.0.0",
    "pydantic>=2.0.0",
    "pytest>=7.0.0",
    "hypothesis>=6.0.0",
    "pyyaml>=6.0.0",
    "typing-extensions>=4.0.0",
    "matplotlib>=3.7.0",
    "scipy>=1.11.0",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "black", 
    "pytest-cov",
]
streamlit = [
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
    "altair>=5.1.0",
    "redis>=5.0.0",
]
sensitivity = [
    "scikit-learn>=1.3.0",
    "seaborn>=0.12.0",
    "sqlalchemy>=2.0.0",
    "joblib>=1.3.0",
]
all = [
    "dca-simulation[streamlit,sensitivity]",
]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
target-version = "py312"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v" 